% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload.R
\name{calcarea}
\alias{calcarea}
\title{calculate sphere volume}
\usage{
calcarea(x, y, type = "square")
}
\arguments{
\item{x}{widght of a tumour in mm.}

\item{y}{height of a tumour in mm.}

\item{type}{either "square" or "circle".}
}
\value{
area in mm2.
}
\description{
This function takes height and width of a tumour and calculates the
area. Some lisences require this parameter to decide if an animal requires
to be culled.
}
