% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload.R
\name{calcsize}
\alias{calcsize}
\title{calculate sphere volume}
\usage{
calcsize(x, y)
}
\arguments{
\item{x}{widght of a tumour in mm.}

\item{y}{height of a tumour in mm.}
}
\value{
volume in mm3.
}
\description{
This function takes height and width of a tumour and calculates the
volume. As the width and height are interchangable in a mice use the highest
value as height.
}
