% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{colourpick}
\alias{colourpick}
\title{Consistent colouring}
\usage{
colourpick(vars, colourtype = "Dark2")
}
\arguments{
\item{vars}{A vector of different experimental conditions.}

\item{colourtype}{One of the RColorBrewer sets. "BW" can be specified if
you desire black and white images. this supports up to four vars.
Default = "Dark2.}
}
\value{
A named vector with hex type colour codes per levels of "vars".
}
\description{
This function returns a vector with colours assigned to each treatment type
for consistent colours over the different plots.
}
