% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataload.R
\name{dataprep}
\alias{dataprep}
\title{create tumour volume matrix}
\usage{
dataprep(file)
}
\arguments{
\item{file}{A table with the first 3 columns indicating sample data:
cage, treatment, mouse/repeat. The remaining columns are dates with the
tumour volume on that day. e.g. 01-01-2020 | 02-01-2020}
}
\value{
A matrix of tumour volumes per mice (rows) for each
 measurement date (columns).
}
\description{
This function reads an input file containing tumour volume data and
returns a matrix with rows = mice and columns = dates of measurement.
}
