% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{endcalc}
\alias{endcalc}
\title{cul date}
\usage{
endcalc(matrix = micematrix)
}
\arguments{
\item{matrix}{matrix of the format: rows = mice, columns = dates,
fill = tumour volume.}
}
\value{
A vector with dates of death per animal, NA if the animal is alive.
}
\description{
This function grabs the last measured tumour size and determines the date
the animal died.
}
