% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{exclude}
\alias{exclude}
\title{Exclude mice}
\usage{
exclude(filledmat, culdat)
}
\arguments{
\item{culdat}{A dataframe which contains the date of death of each mice.
expected format is: 1. Cage, 2. Treatment, 3. Mice ID, 4. Date of death in
format %d/%m/%Y.}

\item{filledmatrix}{matrix of the format: rows = mice, columns = dates,
fill = tumour volume. This function does not allow columns to be skipped.
each day must be represented as a column.}
}
\value{
A TRUE/FALSE vector which indicates if a mice should be excluded or
not.
}
\description{
This function determines which mice to exclude from the study. I does this
by checking if a mice which lack entry data in the final measurement has
got a date of death.
}
