% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{exprun}
\alias{exprun}
\title{experiment time}
\usage{
exprun(start = begin, curdate = Sys.Date())
}
\arguments{
\item{start}{Start date of the animal/experiment. As calculated by startpick}

\item{curdate}{The current date, or the date at which point the
experiment ended.}
}
\value{
A vector of days.
}
\description{
This function calculates how many days have passed for each animal since
they were eligble for the experiment.
}
