% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filldate}
\alias{filldate}
\title{matrix gap fill.}
\usage{
filldate(data, intervaltime, datecolumn = col)
}
\arguments{
\item{data}{matrix of the format: rows = mice, columns = dates,
fill = tumour volume.}

\item{intervaltime}{Vector of the length in days of each interval between
two measurements.}

\item{datecolumn}{A vector of dates of tumour measurement.}
}
\value{
a matrix where every column is a day, if no measurements were made
that day the column is filled with NA.
}
\description{
Before we can plot anything we need to know how the data looks with days
between measurements included, we do this by filling in the missing columns
using this function.
}
