% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{growthindicator}
\alias{growthindicator}
\title{summary of tumour progression}
\usage{
growthindicator(intermatrix = intgrowth, intervaltime = int)
}
\arguments{
\item{intermatrix}{a matrix of tumour volume over intervals, as generated by
growthinterval. The matrix should be mice (rows) x intervals (columns).}

\item{intervaltime}{Vector of the length in days of each interval between
two measurements.}
}
\value{
a data.frame of three columns (growth, remission, stable) with
counts for each mice.
}
\description{
This function summates the amount of days a tumour sees growth, recession or
remains at the same volume over the course of treatment.
}
