% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{growthinterval}
\alias{growthinterval}
\title{growth over intervals}
\usage{
growthinterval(volumematrix = micematrix, datecolumn = col)
}
\arguments{
\item{volumematrix}{matrix of the format: rows = mice, columns = dates,
fill = tumour volume.}

\item{datecolumn}{Vector of dates of measurements, in case they differ from
the column names.}
}
\value{
A matrix of growth of each mice (rows) per intervan (column).
}
\description{
This function calculates the growth of a tumour over each of the measurement
 interval.
}
