% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotdat}
\alias{plotdat}
\title{ggplot data prep}
\usage{
plotdat(data, date = FALSE)
}
\arguments{
\item{data}{a dataframe with 3 columns of identifiers followed by columns of
tumour growth values.}

\item{date}{indicate if the column names of each tumour growth values
are since onset experiment (1-...) or dates (01-01-2020). default = FALSE.}
}
\value{
A named vector with hex type colour codes per levels of "vars".
}
\description{
This function creates a unique ID from the cage and mouse id. "Cage: A1 and
Mouse ID : L1 will create ID A1L1. it will subsequently melt the data to a
long format which can be used by ggplot2.
}
