% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{plotmatrix}
\alias{plotmatrix}
\title{experimental matrix setup}
\usage{
plotmatrix(filledmatrix = micematrix, startdate = first,
  intervaltime = int)
}
\arguments{
\item{filledmatrix}{matrix of the format: rows = mice, columns = dates,
fill = tumour volume. This function does not allow columns to be skipped.
each day must be represented as a column.}

\item{startdate}{Start date of the animal/experiment.
As calculated by startpick}

\item{intervaltime}{Vector of the length in days of each interval between
two measurements.}
}
\value{
A matrix where the first column is day 1 of experiment. column names
 are days since onset experiment, columns are different mice.
}
\description{
This function creates a matrix of the tumour size per date with all data
starting at the experimental start point. e.g., the first column is
experiment day 1 for all mice.
}
