% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvivo.R
\name{rvivo}
\alias{rvivo}
\title{main function of Rvivo}
\usage{
rvivo(volumes = NULL, measures = NULL, cul = NULL, min = 0,
  max = NULL, treatcolours = NULL, colstyle = "Dark2",
  figures = TRUE, axis = "linear", output = "pdf", table = TRUE,
  survival = TRUE, folder = getwd())
}
\arguments{
\item{volumes}{A table with three columns of sample information
(e.g. Cage, treatment, mice tag) and subsequent columns of tumour volumes
in mm3. Column names should be in format Year-month-day, e.g. 2020-01-01.}

\item{measures}{A table with three columns of sample information
(e.g. Cage, treatment, mice tag) and subsequent columns of Height and Width
in side by side columns, both with the date of measurement as column name.}

\item{cul}{A table with three columns of sample information followed with
the date a mice was culled. mice still undergoing treatment or those that
have been excluded from the study should be left blank.}

\item{min}{minimum volume of the tumour before animal can start treatment
(default = 0)}

\item{max}{A value above which tumour entries will be removed from the
analysis.}

\item{treatcolours}{For when you want to use the same colours per treatment
over several images. Input should be a named vector with hex style colour
codes. The names should correspond to the treatments in the data.}

\item{colstyle}{Indicate the type of colour coding that should be assigned
to each treatment group. follows RColorBrewer style. (default = "Dark2")}

\item{figures}{TRUE/FALSE, Indicates if  Rvivo generates plots.
(default = TRUE)}

\item{axis}{"linear" or "log", passes on axis infomation for the plots.}

\item{output}{image extension types: "tiff", "pdf".}

\item{table}{TRUE/FALSE, Inidcates if Rvivo folders a table with results.
(default = TRUE)}

\item{survival}{TRUE/FALSE, inidcates if Rvivo folders Kaplan-meier
survival curves.
 (default = TRUE)}

\item{folder}{The folder Rvivo will ouptut to. (default = Working directory)}
}
\value{
A data.frame of three columns (growth, remission, stable) with
counts for each mice. and depending on table & figures, optionally a
summary file: in_vivo_summary.csv and figures.
}
\description{
This function combines all previous functions to run, it takes an table of
either tumour volumes or measurements. Calculates the start date of the
experiment when mice are above a designated weight. It then calculates,
from this start date, how much the tumour has grown / shrunk since the
starting size. Additionally, it determines progression of the tumour over
the intervals. and how many days of growth, regression or stability the
tumour has shown over the course of the treatment.
}
\details{
provide either volumes or measures, not both.
}
