% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{startpick}
\alias{startpick}
\title{picks experiment start date}
\usage{
startpick(data = micemat, threshold = 90)
}
\arguments{
\item{data}{matrix of the format: rows = mice, columns = dates,
fill = tumour volume.}

\item{threshold}{weight above which the animal is ready for treatment.}
}
\value{
vector of start dates.
}
\description{
This function takes a matrix of tumour volumes per date and returns the
first date measured above a set weight.
}
