% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{survdata}
\alias{survdata}
\title{Mice survival data preparation}
\usage{
survdata(startdate, treatmenttime, culdat)
}
\arguments{
\item{startdate}{Start date of the animal/experiment as calculated by
startpick}

\item{treatmenttime}{how many days have passed for each animal since
they were eligble for the experiment.}

\item{culdat}{A dataframe which contains the date of death of each mice.
expected format is: 1. Cage, 2. Treatment, 3. Mice ID, 4. Date of death in
format %d/%m/%Y.}
}
\value{
a table with the Treatment, survival time and binary survival
indication (1 = death, 0 = alive).
}
\description{
This function prepares the data for a survival analysis. calculating
the survival time of each animal and if the data is censored or not.
}
