% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{survstats}
\alias{survstats}
\title{Calculate survival statistics}
\usage{
survstats(survivaldat = survframe)
}
\arguments{
\item{survivaldat}{A table with treatment, survival time and binary survival
indicator (1 = death, 0 = alive). as generated by survdata.}
}
\value{
a data.table containing the coxph calculated coefficients.
}
\description{
This function applies the Cox proportional hazards regression model and
Kaplan Meier Log Rank statistical test comparing all treatments to the each
other and returns a table.
}
