% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{tumgrowth}
\alias{tumgrowth}
\title{tumour growth}
\usage{
tumgrowth(matrix = micematrix, startdate = begin)
}
\arguments{
\item{matrix}{matrix of the format: rows = mice, columns = dates,
fill = tumour volume.}

\item{startdate}{Start date of the animal/experiment.
As calculated by startpick}
}
\value{
A vector of difference in tumour size.
}
\description{
This function grabs the earliest measured tumour size sinds experiment start
and substracts the last measured tumour size.
}
