% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vivoplot_overall}
\alias{vivoplot_overall}
\title{plot all}
\usage{
vivoplot_overall(data, colours, error = T, dots = T)
}
\arguments{
\item{data}{a long format of the tumour growth data as generated by plotdat.}

\item{colours}{A named vector with hex colours for each treatment variable
as generated by colourpick.}

\item{error}{TRUE/FALSE if error bars should be displayed for each treatment
mean. default = TRUE.}

\item{dots}{TRUE/FALSE if individual measurements should be displayed.
default = TRUE.}
}
\value{
A ggplot showing growth of all treatment groups.
}
\description{
This function creates a ggplot. It displays the growth of all treatment
conditions over time in one graph. It does this by plotting individual
measurements and the mean of each treatment as a line with error bars.
Colours are assigned by Treatment name as the colourpick function generates.
}
