% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vivoplot_treatment}
\alias{vivoplot_treatment}
\title{plot growth}
\usage{
vivoplot_treatment(data, colours = colour, line = T, axis = "linear")
}
\arguments{
\item{data}{a long format of the tumour growth data as generated by plotdat.}

\item{colours}{A named vector with hex colours for each treatment variable
as generated by colourpick.}

\item{line}{TRUE/FALSE If TRUE a line for each animal will be plotted,
otherwise dots will indicate each data point per animal. default = TRUE.}

\item{axis}{"linear" or "log", passes on axis infomation for the plots.
Default = "linear"}
}
\value{
A ggplot of tumour growth per treatment group.
}
\description{
This function creates a ggplot that displays growth of each treatment in a
seperate graph. It plots a line displaying the growth over time for each
individual animal.
Colours are assigned by Treatment name as the colourpick function generates.
}
