% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvivo.R
\name{rvivo}
\alias{rvivo}
\title{main function of Rvivo}
\usage{
rvivo(volumes = NULL, measures = NULL, cul = NULL, minvol = NULL,
  treatcolours = NULL, colstyle = "Dark2", figures = FALSE,
  table = TRUE, output = getwd())
}
\arguments{
\item{volumes}{A table with three columns of sample data
(e.g. Cage, treatment, mice tag) and following columns of tumour volumes
in mm3. Column names should be in format day/month/Year, e.gg.01/01/2020.}

\item{measures}{A table with three columns of sample data
(e.g. Cage, treatment, mice tag) and following columns of Height and Width
in side by side columns, both with the date of measurement as column name.}

\item{minvol}{minimum volume of the tumour before animal can start treatment
(default = 90)}

\item{figures}{TRUE/FALSE, Indicates if  Rvivo generates plots.
(WIP, default = FALSE)}

\item{table}{TRUE/FALSE, Inidcates if Rvivo outputs a table with results.
(default = TRUE)}

\item{output}{The folder Rvivo will ouptut to. (default = Working directory)}
}
\value{
A data.frame of three columns (growth, remission, stable) with
counts for each mice. and depending on table & figures, optionally a
summary file: in_vivo_summary.csv and figures.
}
\description{
This function combines all previous functions to run, it takes an table of
either tumour volumes or measurements. Calculates the start date of the
experiment when mice are above a designated weight. It then calculates,
from this start date, how much the tumour has grown / shrunk since the
starting size. Additionally, it determines progression of the tumour over
the intervals. and how many days of growth, regression or stability the
tumour has shown over the course of the treatment.
}
\details{
provide either volumes or measures, not both.
}
