% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vivoplot_overall}
\alias{vivoplot_overall}
\title{plot all}
\usage{
vivoplot_overall(data, type = "reg", colours = colour,
  axis = "linear")
}
\arguments{
\item{data}{a long format of the tumour growth data as generated by plotdat.}

\item{type}{Should the plots be formatted in percentage growth or
regular values? options: "reg", "per".
Default = "reg"}

\item{colours}{A named vector with hex colours for each treatment variable
as generated by colourpick.}

\item{axis}{"linear" or "log", passes on axis infomation for the plots.
Default = "linear"}
}
\value{
A ggplot showing growth of all treatment groups.
}
\description{
This function creates a ggplot. It displays the growth of all treatment
conditions over time in one graph. It does this by plotting individual
measurements and the mean of each treatment as a line with error bars.
Colours are assigned by Treatment name as the colourpick function generates.
}
